+++ rust/src/day5.rs
@@ -1,7 +1,22 @@
+//! Day 5: Sunny with a Chance of Asteroids
+//!
+//! # Problem Description
+//!
+//! Second day dealing with the Intcode machine. This time we had to
+//! expand it with 4 additional operations and an extra mode for
+//! operation parameters. The actual day-specific task was to run a
+//! test program to make sure your implementation was fully working
+//! properly.
+//!
+//! #Implementation Details
+//!
+//! This is when I moved the intcode stuff to a separate model which
+//! can be day-agnostic. See [intcode] for more information.
+
 use super::intcode;
 
-pub fn part1(input: &str) -> isize {
-    let mut codes: Vec<isize> = input
+pub fn part1(input: &str) -> i64 {
+    let mut codes: Vec<i64> = input
         .trim()
         .split(",")
         .map(|s| s.parse().unwrap())
